LayoutTests/svg/custom/loadevents-async-expected.txt

This test ensures that tree mutation on the load doesn't break consistency.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS invoked is false
PASS invoked is true
PASS successfullyParsed is true

TEST COMPLETE
